requirements.txt
# Evaluation AI Dependencies

# Core
langfuse>=0.8.0
openai>=1.0.0
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data
numpy>=1.24.0
